Allow ConversionJob to process 100 usernames
HTTPProfileRepository#findProfilesByCriteria supports 50 usernames per
results page and is terminated by an empty results page, i.e., querying
with 50 usernames will hit page one for 50 results, and stop after it
receives the second (empty) page.

This makes it immediately obvious that we should process at least 50
usernames instead of 1 per job, and in testing better results were
achieved with 100 usernames per ConversionJob (3 queries for 100
usernames, as opposed to 2 queries for 50).

The speed may still be increased with a higher number of usernames per
job, but I've only tested up to 100.